move exporter libs
